move test closer to what's being tested + simplify
