+++ simapp/app_test.go
@@ -12,9 +12,7 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/require"
 
-	"cosmossdk.io/core/address"
 	"cosmossdk.io/core/appmodule"
-	"cosmossdk.io/depinject"
 	"cosmossdk.io/log"
 	"cosmossdk.io/x/accounts"
 	"cosmossdk.io/x/auth"
@@ -297,62 +295,4 @@ func TestProtoAnnotations(t *testing.T) {
 	require.NoError(t, err)
 	err = msgservice.ValidateProtoAnnotations(r)
 	require.NoError(t, err)
-}
-
-var _ address.Codec = (*customAddressCodec)(nil)
-
-type customAddressCodec struct{}
-
-func (c customAddressCodec) StringToBytes(text string) ([]byte, error) {
-	return []byte(text), nil
-}
-
-func (c customAddressCodec) BytesToString(bz []byte) (string, error) {
-	return string(bz), nil
-}
-
-func TestAddressCodecFactory(t *testing.T) {
-	var addrCodec address.Codec
-	var valAddressCodec address.ValidatorAddressCodec
-	var consAddressCodec address.ConsensusAddressCodec
-
-	err := depinject.Inject(
-		depinject.Configs(
-			AppConfig(),
-			depinject.Supply(log.NewNopLogger()),
-		),
-		&addrCodec, &valAddressCodec, &consAddressCodec)
-	require.NoError(t, err)
-	require.NotNil(t, addrCodec)
-	_, ok := addrCodec.(customAddressCodec)
-	require.False(t, ok)
-	require.NotNil(t, valAddressCodec)
-	_, ok = valAddressCodec.(customAddressCodec)
-	require.False(t, ok)
-	require.NotNil(t, consAddressCodec)
-	_, ok = consAddressCodec.(customAddressCodec)
-	require.False(t, ok)
-
-	// Set the address codec to the custom one
-	err = depinject.Inject(
-		depinject.Configs(
-			AppConfig(),
-			depinject.Supply(
-				log.NewNopLogger(),
-				func() address.Codec { return customAddressCodec{} },
-				func() address.ValidatorAddressCodec { return customAddressCodec{} },
-				func() address.ConsensusAddressCodec { return customAddressCodec{} },
-			),
-		),
-		&addrCodec, &valAddressCodec, &consAddressCodec)
-	require.NoError(t, err)
-	require.NotNil(t, addrCodec)
-	_, ok = addrCodec.(customAddressCodec)
-	require.True(t, ok)
-	require.NotNil(t, valAddressCodec)
-	_, ok = valAddressCodec.(customAddressCodec)
-	require.True(t, ok)
-	require.NotNil(t, consAddressCodec)
-	_, ok = consAddressCodec.(customAddressCodec)
-	require.True(t, ok)
 }
